requirements.txt
python-decouple
langchain-openai
langchain_community
pydantic
pdfminer.six